Split react-redux wrapper into few layers
