+++ src/main/scala/package.scala
@@ -1,9 +1,6 @@
 package eldis
 
 import scalajs.js
-import japgolly.scalajs.react._
-import org.scalajs.dom
-import dom.raw.Element
 
 /**
  * The redux facade object.
@@ -70,52 +67,4 @@ package object redux {
   @inline def wrapAction[A](a: A) = Redux.wrapAction(a)
 
   @inline def wrapAction[A](a: Future[A]) = Redux.wrapAction(a)
-}
-
-package redux {
-  /**
-   * React redux facade.
-   */
-  package object react {
-
-    /**
-     * The react component that provides the store injection in to the virtual DOM.
-     *
-     * See [[https://github.com/reactjs/react-redux/blob/master/docs/api.md#provider-store the react-redux documentation]]
-     * for detailed description.
-     */
-    val Provider = ReactRedux.Japgolly.Provider
-
-    /** The function that maps the state and the dispatcher function to the component's properties */
-    type Connector[S, A, P] = ReactRedux.Japgolly.Connector[S, A, P]
-
-    type ConnectedComponentFactory[Props, State, +Backend, +Node <: TopNode] = ReactRedux.Japgolly.ConnectedComponentFactory[Props, State, Backend, Node]
-
-    /**
-     * Creates the connected to state component factory.
-     *
-     * @param connector  The function that maps state and dispatcher function to component's properties
-     * @param cls        The component's class
-     */
-    @inline def connect[S, A, P, S1, B](connector: Connector[S, A, P], cls: ReactClass[P, S1, B, Element]): ConnectedComponentFactory[P, S1, B, Element] =
-      ReactRedux.Japgolly.connect(connector, cls)
-
-    /**
-     * Connects stateless component to the global state
-     *
-     * @param connector  The function that maps state and dispatcher function to component's properties
-     * @param comp       The component
-     */
-    @inline def connect[S, A, P](connector: Connector[S, A, P], comp: FunctionalComponent[P]): FunctionalComponent[P] =
-      ReactRedux.Japgolly.connect(connector, comp)
-
-    /**
-     * Connects stateless component with children to the global state
-     *
-     * @param connector  The function that maps state and dispatcher function to component's properties
-     * @param comp       The component
-     */
-    @inline def connect[S, A, P](connector: Connector[S, A, P], comp: FunctionalComponent.WithChildren[P]): FunctionalComponent.WithChildren[P] =
-      ReactRedux.Japgolly.connect(connector, comp)
-  }
 }
